CBF/requirements/base.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.3
cmsplugin-filer==1.1.3
decorator==4.0.11
dj-database-url==0.4.2
Django==1.10.7
django-appconf==1.0.2
django-autoslug==1.9.3
django-classy-tags==0.8.0
django-cms==3.4.4
django-compressor==2.1.1
django-extensions==1.8.1
django-filer==1.2.7
django-filter==1.0.4
django-fontawesome==0.3.1
django-formtools==2.0
django-model-utils==3.0.0
django-mptt==0.8.7
django-polymorphic==1.0.2
django-sekizai==0.10.0
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.2.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.0
djangocms-installer==0.9.7
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
easy-thumbnails==2.4.1
facebook-sdk==2.0.0
flake8==3.3.0
future==0.16.0
html5lib==0.9999999
idna==2.5
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==4.0.2
mccabe==0.6.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
rcssmin==1.0.6
redis==2.10.5
requests==2.17.3
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
tweepy==3.5.0
tzlocal==1.4
Unidecode==0.4.20
urllib3==1.21.1
vine==1.1.3
wcwidth==0.1.7
